rebased with recent PR, updated api calls in tests
